+++ plugins/auth/mariadb_cleartext.c
@@ -48,7 +48,7 @@ static int clear_password_auth_client(MYSQL_PLUGIN_VIO *vio, MYSQL *mysql)
     return CR_ERROR;
 
   /* write password including terminating zero character */
-  return vio->write_packet(vio, (const unsigned char *) mysql->passwd, strlen(mysql->passwd) + 1) ?
+  return vio->write_packet(vio, (const unsigned char *) mysql->passwd, (int)strlen(mysql->passwd) + 1) ?
          CR_ERROR : CR_OK;
 }
 /* }}} */
